Struct nova_notes::notebook::db::EmbedStore

pub struct EmbedStore {
    embedding_model: TextEmbedding,
    db_conn: Connection,
    table: Table,
}

Fields
embedding_model: TextEmbeddingdb_conn: Connectiontable: Table
Implementations

impl EmbedStore

pub async fn new( embedding_model: TextEmbedding ) -> Result<EmbedStore, EmbedStoreError>

pub async fn add( &self, text: Vec<String>, alt_ids: Vec<String> ) -> Result<(), EmbedStoreError>

pub async fn record_count(&self) -> Result<usize, EmbedStoreError>

pub async fn search( &self, search_text: &str, limit: Option<usize> ) -> Result<Vec<Document>, EmbedStoreError>

async fn execute_query( &self, query: Option<Vec<f32>>, filter: Option<String>, limit: Option<usize> ) -> Result<Vec<Document>, EmbedStoreError>

pub async fn get(&self, id: &str) -> Result<Option<Document>, EmbedStoreError>

pub async fn get_all(&self) -> Result<Vec<Document>, EmbedStoreError>

pub async fn delete<T: Display>(&self, id: T) -> Result<(), EmbedStoreError>

pub async fn update( &self, id: &str, text: Vec<String> ) -> Result<(), EmbedStoreError>

pub async fn create_index( &self, num_partitions: Option<u32> ) -> Result<(), EmbedStoreError>

Creates an index on a given field.

fn record_to_document(&self, record_batch: Vec<RecordBatch>) -> Vec<Document>

async fn init_db_conn() -> Result<Connection, EmbedStoreError>

fn create_embeddings( &self, documents: &[String] ) -> Result<Vec<Embedding>, EmbedStoreError>

fn wrap_in_option<T>(&self, source: Vec<Vec<T>>) -> Vec<Option<Vec<Option<T>>>>

Transforms a 2D vector into a 2D vector where each element is wrapped in an Option.

This function takes a 2D vector source as input and returns a new 2D vector where each element is wrapped in an Option. The outer vector is also wrapped in an Option. This is useful when you want to represent the absence of data in your vector.

Arguments

source - A 2D vector that will be transformed.

Returns

A 2D vector where each element is wrapped in an Option, and the outer vector is also wrapped in an Option.

Example

let source = vec![vec![1, 2, 3], vec![4, 5, 6]];
let result = wrap_in_option(source);
assert_eq!(result, vec![Some(vec![Some(1), Some(2), Some(3)]), Some(vec![Some(4), Some(5), Some(6)])]);

fn create_record_batch( &self, embeddings: Vec<Vec<f32>>, text: Vec<String>, alt_ids: Vec<String>, schema: Arc<Schema> ) -> Vec<RecordBatch>

Creates a record batch from a list of embeddings and a correlated list of original text.

fn generate_schema(dimensions_count: usize) -> Arc<Schema>

async fn get_or_create_table( db_conn: &Connection, table_name: &str ) -> Result<Table>

async fn create_empty_table( &self, table_name: &str, dimensions_count: usize ) -> Result<Table>

Creates an empty table with a schema.